fix: Split XKB code out of XWindowsUtil to allow building without X11
The xorgproto headers are still required for EI, but nothing more. Note
that the GUI hasn't needed X11 for a while.
