+++ src/gui/CMakeLists.txt
@@ -217,8 +217,8 @@ if (BUILD_CARBON)
     find_library(APPSERVICES_LIB ApplicationServices)
     target_link_libraries(input-leap ${APPSERVICES_LIB})
 endif()
-if (BUILD_XWINDOWS)
-    target_link_libraries (input-leap dns_sd X11 Threads::Threads)
+if (BUILD_XWINDOWS OR BUILD_LIBEI)
+    target_link_libraries (input-leap dns_sd Threads::Threads)
 endif()
 
 target_link_libraries (input-leap common)
